feat: refactored dashboard chart
