+++ src/ducks/Stablecoins/StablecoinsMarketCap/StablecoinsMarketCap.js
@@ -1,13 +1,10 @@
 import React, { useEffect, useMemo, useState } from 'react'
 import cx from 'classnames'
-import withSizes from 'react-sizes'
 import Button from '@santiment-network/ui/Button'
 import DashboardChartHeader, {
   DashboardIntervals
 } from '../../../components/DashboardMetricChart/DashboardChartHeader/DashboardChartHeader'
 import DashboardChartMetrics from '../../../components/DashboardMetricChart/DashboardChartMetrics/DashboardChartMetrics'
-import Chart from '../../Chart'
-import { useMetricCategories } from '../../Chart/Synchronizer'
 import {
   MARKET_CAP_YEAR_INTERVAL,
   STABLE_COINS_MARKETCAP_INTERVALS,
@@ -16,27 +13,12 @@ import {
 } from './utils'
 import { useStablecoinsTimeseries } from './hooks'
 import { DesktopOnly, MobileOnly } from '../../../components/Responsive'
-import { mapSizesToProps } from '../../../utils/withSizes'
 import SharedAxisToggle from '../../Studio/Chart/SharedAxisToggle'
-import { useDomainGroups, useAxesMetricsKey } from '../../Chart/hooks'
 import { formIntervalSettings } from '../../SANCharts/IntervalSelector'
+import DashboardMetricChartWrapper from '../../../components/DashboardMetricChart/DashboardMetricChartWrapper'
 import styles from './StablecoinsMarketCap.module.scss'
 
-const CHART_HEIGHT = 400
-const CHART_PADDING_DESKTOP = {
-  top: 32,
-  right: 74,
-  bottom: 58,
-  left: 24
-}
-const CHART_PADDING_MOBILE = {
-  top: 0,
-  right: 0,
-  bottom: 0,
-  left: 0
-}
-
-const StablecoinsMarketCap = ({ isDesktop, className }) => {
+const StablecoinsMarketCap = ({ className }) => {
   const [interval, setInterval] = useState(MARKET_CAP_YEAR_INTERVAL)
   const [disabledAssets, setDisabledAsset] = useState({})
   const [isDomainGroupingActive, setIsDomainGroupingActive] = useState(true)
@@ -62,15 +44,6 @@ const StablecoinsMarketCap = ({ isDesktop, className }) => {
     () => metrics.filter(({ slug }) => !disabledAssets[slug]),
     [metrics, disabledAssets]
   )
-
-  const categories = useMetricCategories(filteredMetrics)
-
-  const axesMetricKeys = useAxesMetricsKey(
-    filteredMetrics,
-    isDomainGroupingActive
-  ).slice(0, 1)
-
-  const domainGroups = useDomainGroups(filteredMetrics)
 
   return (
     <div className={cx(styles.container, className)}>
@@ -117,22 +90,13 @@ const StablecoinsMarketCap = ({ isDesktop, className }) => {
         />
       </DesktopOnly>
 
-      <Chart
-        {...settings}
-        {...categories}
-        data={data}
-        chartHeight={CHART_HEIGHT}
+      <DashboardMetricChartWrapper
         metrics={filteredMetrics}
-        isCartesianGridActive={false}
-        hideWatermark
-        hideBrush
-        chartPadding={isDesktop ? CHART_PADDING_DESKTOP : CHART_PADDING_MOBILE}
-        resizeDependencies={[]}
+        data={data}
+        settings={settings}
         MetricColor={StablecoinColor}
-        tooltipKey={axesMetricKeys[0]}
-        axesMetricKeys={axesMetricKeys}
-        domainGroups={isDomainGroupingActive ? domainGroups : undefined}
-        isLoading={loadings.length > 0}
+        isDomainGroupingActive={isDomainGroupingActive}
+        loadings={loadings}
       />
 
       <MobileOnly>
@@ -152,4 +116,4 @@ const StablecoinsMarketCap = ({ isDesktop, className }) => {
   )
 }
 
-export default withSizes(mapSizesToProps)(StablecoinsMarketCap)
+export default StablecoinsMarketCap
